feat: polish the test helper function
