+++ tests/integration/monitoring/test_request_size.py
@@ -1,10 +1,10 @@
-from typing import Dict, List
-
 import pytest
 import requests as req
 from docarray import DocumentArray
 
 from jina import Executor, Flow, requests
+
+from . import get_metric_values
 
 
 @pytest.fixture()
@@ -68,27 +68,6 @@ def test_requests_size(port_generator, executor):
 
         assert measured_request_bytes_sum > measured_request_bytes_sum_init
         assert measured_request_bytes_send_sum > measured_request_bytes_send_sum_init
-
-
-def get_metric_values(raw_metrics: str) -> Dict[str, float]:
-    """
-    get the value of a metric from the prometheus endpoint
-    :param raw_metrics: raw string coming from scrapping the http prometheus endpoint
-    :return: Dictionary which full metrics name as key and the corresponding value
-    """
-    metrics = dict()
-
-    for line in raw_metrics.split('\n'):
-        if not line.startswith('#') and ' ' in line:
-            line_split = line.split(' ')
-
-            metric_name = ''.join(line_split[0:-1])
-
-            metric_value = float(line_split[-1])
-
-            metrics[metric_name] = metric_value
-
-    return metrics
 
 
 def test_request_size_increasing(port_generator, executor):
